+++ graf2d/win32gdk/gdk/src/gdk/gdkfont.h
@@ -30,7 +30,7 @@ extern "C" {
    GdkFont *gdk_fontset_load(const gchar * fontset_name);
    GdkFont *gdk_font_ref(GdkFont * font);
    void gdk_font_unref(GdkFont * font);
-   gint gdk_font_id(const GdkFont * font);
+   glong gdk_font_id(const GdkFont * font);
    gboolean gdk_font_equal(const GdkFont * fonta, const GdkFont * fontb);
 
    gint gdk_string_width(GdkFont * font, const gchar * string);
